requirements.txt
pyTelegramBotAPI
Pillow
pillow-heif
reportlab
Flask